api_proveedores/requirements.txt
asgiref==3.2.7
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-cors-headers==3.5.0
django-countries==7.2.1
django-import-export==2.6.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
drf-api-tracking==1.8.0
drf-yasg==1.17.1
future==0.18.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
inflection==0.4.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.0
ordered-set==4.0.2
packaging==20.4
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyuca==1.2
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==2.0.1
PyYAML==5.3.1
google-api-python-client
google-auth-oauthlib
